Add TrackPlayer struct
TrackPlayer offers some simple audio controls over the currently playing track. This addition required the ReadCloser returned as a part of Track to be changed to a custom ReadSeekCloser due to the underlying beep library requiring a Seeker to change position in a track. This means tracks must be fully download before this can be played; otherwise, there is no way to effectively seek around the track. As a result, this change introduces some playback lag which is undesirable, and future work should investigte how to revert back to a streaming interface while still using beep

ReadSeekerCloser -> ReadSeekCloser